requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.15.0
pandas>=1.3.0
scikit-learn>=0.24.2
matplotlib>=3.4.3
numpy>=1.20.0
tqdm>=4.62.0
pillow>=8.3.1
accelerate>=0.5.1 